# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Air Ticket Booking API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/bookings/": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Create Booking",
        "operationId": "create_booking_api_bookings__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/bookings/me": {
      "get": {
        "tags": [
          "bookings"
        ],
        "summary": "List My Bookings",
        "operationId": "list_my_bookings_api_bookings_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "type": "array",
                  "title": "Response List My Bookings Api Bookings Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/bookings/{booking_id}": {
      "get": {
        "tags": [
          "bookings"
        ],
        "summary": "Get My Booking Detail",
        "operationId": "get_my_booking_detail_api_bookings__booking_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/me": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Me",
        "operationId": "admin_me_api_admin_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/dashboard": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Dashboard",
        "operationId": "admin_dashboard_api_admin_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDashboard"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/overrides": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Flight Overrides",
        "operationId": "list_flight_overrides_api_admin_overrides_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlightOverrideResponse"
                  },
                  "type": "array",
                  "title": "Response List Flight Overrides Api Admin Overrides Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Flight Override",
        "operationId": "create_flight_override_api_admin_overrides_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightOverrideCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightOverrideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/overrides/{override_id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Flight Override",
        "operationId": "update_flight_override_api_admin_overrides__override_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "override_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Override Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightOverrideUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightOverrideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Flight Override",
        "operationId": "delete_flight_override_api_admin_overrides__override_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "override_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Override Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bookings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Bookings",
        "operationId": "list_all_bookings_api_admin_bookings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "title": "Response List All Bookings Api Admin Bookings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bookings/{booking_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Booking Detail",
        "operationId": "get_booking_detail_api_admin_bookings__booking_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Booking Status",
        "operationId": "update_booking_status_api_admin_bookings__booking_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bookings/{booking_id}/payment-status": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Payment Status",
        "operationId": "update_payment_status_api_admin_bookings__booking_id__payment_status_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bookings/{booking_id}/upload-ticket": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Ticket",
        "operationId": "upload_ticket_api_admin_bookings__booking_id__upload_ticket_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bookings/auto-cancel": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Trigger Auto Cancel",
        "operationId": "trigger_auto_cancel_api_admin_bookings_auto_cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expire_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Expire Minutes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bookings/{booking_id}/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Booking Audit",
        "operationId": "get_booking_audit_api_admin_bookings__booking_id__audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminDashboard": {
        "properties": {
          "financial": {
            "$ref": "#/components/schemas/DashboardFinancial"
          },
          "operational": {
            "$ref": "#/components/schemas/DashboardOperational"
          },
          "today": {
            "$ref": "#/components/schemas/DashboardToday"
          }
        },
        "type": "object",
        "required": [
          "financial",
          "operational",
          "today"
        ],
        "title": "AdminDashboard"
      },
...
      "BookingCreate": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "adults": {
            "type": "integer",
            "maximum": 9.0,
            "exclusiveMinimum": 0.0,
            "title": "Adults"
          },
          "bundle_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Key"
          },
          "flight_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Flight Snapshot"
          },
          "final_price_usd": {
            "type": "number",
            "title": "Final Price Usd"
          },
          "final_price_mmk": {
            "type": "number",
            "title": "Final Price Mmk"
          }
        },
        "type": "object",
        "required": [
          "type",
          "adults",
          "flight_snapshot",
          "final_price_usd",
          "final_price_mmk"
        ],
        "title": "BookingCreate"
      },
      "BookingOut": {
        "properties": {
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id"
          },
          "booking_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Code"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "adults": {
            "type": "integer",
            "title": "Adults"
          },
          "bundle_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Key"
          },
          "flight_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Flight Snapshot"
          },
          "final_price_usd": {
            "type": "number",
            "title": "Final Price Usd"
          },
          "final_price_mmk": {
            "type": "number",
            "title": "Final Price Mmk"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "passengers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PassengerOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passengers"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "booking_code",
          "type",
          "adults",
          "bundle_key",
          "flight_snapshot",
          "final_price_usd",
          "final_price_mmk",
          "status",
          "payment_status"
        ],
        "title": "BookingOut"
      },
...
      "FlightOverrideCreate": {
        "properties": {
          "airline_code": {
            "type": "string",
            "title": "Airline Code"
          },
          "flight_number": {
            "type": "string",
            "title": "Flight Number"
          },
          "departure_date": {
            "type": "string",
            "format": "date",
            "title": "Departure Date"
          },
          "override_price_usd": {
            "type": "number",
            "title": "Override Price Usd"
          }
        },
        "type": "object",
        "required": [
          "airline_code",
          "flight_number",
          "departure_date",
          "override_price_usd"
        ],
        "title": "FlightOverrideCreate"
      },
      "FlightOverrideResponse": {
        "properties": {
          "airline_code": {
            "type": "string",
            "title": "Airline Code"
          },
          "flight_number": {
            "type": "string",
            "title": "Flight Number"
          },
          "departure_date": {
            "type": "string",
            "format": "date",
            "title": "Departure Date"
          },
          "override_price_usd": {
            "type": "number",
            "title": "Override Price Usd"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "airline_code",
          "flight_number",
          "departure_date",
          "override_price_usd",
          "id"
        ],
        "title": "FlightOverrideResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}